requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.115.6
fastapi-cli==0.0.7
h11==0.16.0
httptools==0.6.4
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
pika==1.3.2
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.0.1
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.3
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.34.0
watchfiles==1.1.0
websockets==15.0.1
pyjwt
python-multipart
passlib[argon2]
